+++ src/packages/checkboxgroup/demo.vue
@@ -1,11 +1,14 @@
 <template>
     <div class="demo-list">
-        <h4>基本用法</h4>
+        <h4>基本用法(2s后动态更新)</h4>
         <div>
             <nut-cell>
-                <span slot="title"><nut-checkboxgroup  :checkBoxData="data1" v-model="group1"></nut-checkboxgroup></span>
+                <span slot="title"><nut-checkboxgroup ref="checkboxGroup" :checkBoxData="data1" v-model="group1"></nut-checkboxgroup></span>
             </nut-cell>
             <p>{{group1}}</p>
+            <nut-button small @click="checkAll(true)">全选</nut-button>
+            <nut-button small @click="checkAll(false)">取消全选</nut-button>
+            <nut-button small @click="checkAll()">反选</nut-button>
         </div>
 
         <h4>禁用状态</h4>
@@ -34,7 +37,7 @@
         <h4>禁用动效</h4>
         <div>
             <nut-cell>
-                <span slot="title"><nut-checkboxgroup  :checkBoxData="['选项1','选项2']" v-model="group6"  :animation="false" ></nut-checkboxgroup></span>
+                <span slot="title"><nut-checkboxgroup  :checkBoxData="data33" v-model="group6"  :animation="false" ></nut-checkboxgroup></span>
 
             </nut-cell>
             <p>选择状态：{{group6}}</p>
@@ -121,13 +124,20 @@ export default {
             ],
         };
     },
+    mounted(){
+        setTimeout(() => {
+            this.group1.push('选项B')
+        }, 2000);
+    },
     methods: {
         changeEvt(val,label,e){
             console.log(0,val,label,e)
             alert('已选值:['+ val+']，当前选择值：'+ label);
            
         },
-        
+        checkAll(state){
+            this.$refs.checkboxGroup.toggleAll(state);
+        }
     }
 };
 </script>