+++ src/packages/scroller/vertical-scroll.vue
@@ -56,12 +56,18 @@ export default {
             type: Number,
             default: 100
         },
+        propsTime: {
+            type:Number,
+            default: 0
+        }
     },
     watch: {
         'isLoading': function(status) {
             if (!status && this.realMove === 0) {
                 clearTimeout(this.timer);
-                this.setTransform(this.realMove, 'end', 0);
+                this.timer = setTimeout(() => {
+                    this.setTransform(this.realMove, 'end', null);
+                }, this.propsTime);
             }
         },
         'isUnMore': function() {
@@ -140,12 +146,12 @@ export default {
                         // }, time / 2);
                     }
                 }
-                if (updateMove == 50 && !this.isLoading) {
-                    clearTimeout(this.timer);
-                    this.timer = setTimeout(() => {
-                        this.setTransform(this.realMove, 'end', null);
-                    }, 3000);
-                }
+                // if (updateMove == 50 && !this.isLoading) {
+                //     clearTimeout(this.timer);
+                //     this.timer = setTimeout(() => {
+                //         this.setTransform(this.realMove, 'end', null);
+                //     }, 3000);
+                // }
                 this.setTransform(updateMove, type, time)
             } else {
                 if (updateMove > 0 && updateMove > this.stretch) {