+++ src/packages/uploader/uploader.vue
@@ -43,6 +43,12 @@ export default {
         return {};
       }
     },
+    headers:{
+      type: Object,
+      default() {
+        return {};
+      }
+    },
     changeEvtCallback: {
       type: Function
     },
@@ -65,6 +71,10 @@ export default {
     limitError: {
       type: String,
       default: ''
+    },
+    withCredentials: {
+      type: Boolean,
+      default: false
     }
   },
   data() {
@@ -77,12 +87,14 @@ export default {
         $el: {},
         url: this.url, //图片上传地址
         formData: null,
+        headers: {}, //自定义headers
         isPreview: this.isPreview, //是否开启本地预览
         previewData: null,
         maxSize: this.maxSize, //允许上传的文件最大字节
         acceptType: this.acceptType, //允许上传的文件类型
         xhrState: this.xhrState,
         clearInput: this.clearInput,
+        withCredentials: this.withCredentials,//支持发送 cookie 凭证信息
         xmlError: this.xmlError || this.nutTranslate('lang.uploader.xmlError'),
         typeError: this.typeError || this.nutTranslate('lang.uploader.typeError'),
         limitError: this.limitError || this.nutTranslate('lang.uploader.limitError'),
@@ -139,6 +151,7 @@ export default {
         formData.append(key, this.attach[key]);
       }
       opt.formData = formData;
+      opt.headers = this.headers || {};
       opt.showMsgFn = msg => {
         this.$emit("showMsg", msg);
       };