+++ src/utils/raf.js
@@ -1,12 +1,14 @@
 function requestAniFrame() {
-    return (
-        window.requestAnimationFrame ||
-        window.webkitRequestAnimationFrame ||
-        window.mozRequestAnimationFrame ||
-        function (callback) {
-            window.setTimeout(callback, 1000 / 60);
-        }
-    );
+    if(typeof window !== 'undefined'){
+        return (window.requestAnimationFrame || window.webkitRequestAnimationFrame || window.mozRequestAnimationFrame || 
+            function (callback) {
+                window.setTimeout(callback, 1000 / 60);
+            });
+    }else{
+        return function (callback) {
+                    setTimeout(callback, 1000 / 60);
+               };
+    }
 }
 
 export default requestAniFrame();